Separated _signaltools_kernels into _convolve_kernels.py and _correlate_kernels.py
